Quiz stats - question banks support
We were using the total participant count for the quiz to calculate
certain response ratios for each question. This patch changes that
behavior to use the number of students who were actually presented with
each question.

There might still be issues with the DI chart [1] but I can not verify
that.

Closes CNVS-17148

TEST PLAN
---- ----

  - Create a quiz with:
    1) 2 normal questions
    2) a quiz group linked to a bank with 1 question pick
    3) let it allow multiple attempts
  - Take it by a number of students so that you get different questions
    pulled out of the bank.
    + After each attempt, refresh the stats page and verify that the
      question stats for the random question you just received is
      correct in that:
        1) The "Attempts: X out of Y" should read like this: X would be
           the number of students who took this question and left any
           NON-EMPTY response, where Y is the number of students who
           were presented with that question altogether, regardless of
           the type of response they left.
        2) The charts for that relevant question are correct[1].
  - If you re-take the quiz by a student who received a unique question
    (i.e, no one else has received it) and you do NOT get that question
    on the subsequent attempt, the stats page will no longer display
    that question since it has no data anymore

[1] The discrimination index chart may not work as expected because of
the way it's calculated; it only looks at the data for the *first*
submitted attempt, and it requires a hefty number of responses. So,
you need to get really lucky that you receive a MC/TF question *on the
first attempt* by a number of students. I never got it to work, YMMV.

Change-Id: I5238b98cfb6ec92461bbcddbf67ef88fc9b37f1d
Reviewed-on: https://gerrit.instructure.com/46007
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Trevor deHaan <tdehaan@instructure.com>