fix issue with zip files not uploading
fixes CNVS-17064

If you drag/drop an image and a .zip file the zip file won't get
uploaded. This was because we were trying to upload the zip before we
asked the user if they wanted to unzip it. This was fixed with a
conditional.

Test Plan

Given new files is enabled
And you are able to upload files
And you have no current files in your current folder uploaded
When you multiselect an image and .zip file from your desktop
And you dragg those files into the file browser
Then both the image and the .zip should be queued for upload
And you should be asked if you want to unzip the .zip file
And after you've made your selection both files should be uploaded

Change-Id: I509c7971154e9cf878ec6f9117689b9a9918c220
Reviewed-on: https://gerrit.instructure.com/46514
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>