Strip nil or "" values from account settings hash
To save space in the DB, we don't need to serialize
account settings values that are either the empty
string or nil.  That just saves a bunch of YAML to
the db that is unnecessary.

I don't think a DB migration that goes through and
cleans everything up is necessary since messing
with data at rest is pointless. If/when the settings
is saved next time it will get pruned.

Change-Id: I0c76267c65a1ab350e5f093862484d6839cbcede
Reviewed-on: https://gerrit.instructure.com/48357
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>