Refactors Add Module Item for Files
Converts the "Add File" portion of the "Add Module Item" dialog
into a React component.  It also sets up a Folder and File store
for interacting with the API.  It also modifies the behavior
so that folders that do not contain any files are not shown in
the select box.

fixes CNVS-13036
closes CNVS-18302

Test Plan:
  - Go to the modules page in a course that has some files.
  - Add a file to a module via the file menu.
  - Create 200+ folders with at least one file in each.
    You can import and expand this zip to make this a lot easier:
     http://goo.gl/lNtfx5
  - All files should be visible.
  - Create another folder, with no files within it.
  - That folder should not be visible.

Change-Id: I93e520135acb66bd821b3d19cc387eec59b347c7
Reviewed-on: https://gerrit.instructure.com/49028
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>