Add quiz version number to API response
From a support standpoint, the only way to check quiz versions is to
attempt the quiz as a student, and look at the version number attached
to the submission.  It would be beneficial to have the current version
attached to the quiz object, so these submissions become unnecessary.

Closes CNVS-19001

Test plan:
- Create a quiz in a course.
- Hit /api/v1/courses/:course_id/quizzes/:id
- Verify the response includes a 'version_number' attribute

Change-Id: I2f3550fb9041673f4bbffebc810e1fdc728439ea
Reviewed-on: https://gerrit.instructure.com/49680
Tested-by: Jenkins
Reviewed-by: Ryan Taylor <rtaylor@instructure.com>
QA-Review: Robert Lamb <rlamb@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>