fix focus order ko user on people page
fixes CNVS-19216

When you are only using your keyboard to navigate on the peopls tab, the
tab order was wrong. This fixes that.

Test Plan
Given I'm on the people's tab in a course
And I have created a "group set"
And within that groupset I've edited it to "allow self sign-up"
And I'm only using a keyboard to navigate
When I use tab to navigate to the "+ Group" button"
Then I should first be focused on the "Self signup is enabled for these
groups" help icon
And when I hit tab again
Then I should focused onto the "+ Group" button

Change-Id: I450920987c75b51b2a7bf108d74cf731a9ea2167
Reviewed-on: https://gerrit.instructure.com/50412
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Jon Willesen <jonw@instructure.com>