grades being erased by csv upload are highlighted
For grades that are chaning during csv upload,
the original grade and the new grade will be
displayed. New grades that are lower than the
current grade will be highlighted in red. If
an original grade was logged and a new grade
is not present, it will also be highlighted
red.

fixes CNVS-5197

scope
- Making sure the grades are highlighted in the
  correct scenario.

  Scenarios:
    - When a student's current grade is higher
      then the new grade

    - When a student's current grade has been
      logged, and the csv that's being uploaded
      doesn't have a grade for the student.

- Making sure the behavior of the new uploader
  is consistent with the old csv uploader. s

Change-Id: Id93413b07387f8caac184b3653c5adee282475e8
Reviewed-on: https://gerrit.instructure.com/50465
Tested-by: Jenkins
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Product-Review: Dylan Ross <dross@instructure.com>