Make calendar event editing datepicker accessible
closes CNVS-19312
closes CNVS-19313

refs CNVS-15543

see the description and comments of CNVS-15543 for full narrative

This applies a screenreader-only label to the date picker used to
edit calendar events in a modal window,
and also adds an on-focus tooltip for KB-only users that gives format
information (the same info for both).

Also provides the first cut at time formats for just-time pickers,
which that modal also makes use of.

ALSO does all this^ for the date/time fields on the
"editCalendarFull" page which is what you go to
when you click "More Options" from the modal overlay

TEST PLAN:
- go to the calendar (Agenda view is fine)
- edit any event in the modal overlay
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit in with no additional context. (should be
  date only, no time)
- do the same for the time fields below that on the event
- tab through to the same fields (date and time in that modal)
  and prove that you see
  a little tooltip telling you what format to enter the date or time in.
- click "More Options"
- Do steps 3-5 again for the fields on the full calendar event editor

Change-Id: I41df3816a47c2a2a89efceff5700ab4d9fa77b63
Reviewed-on: https://gerrit.instructure.com/50537
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins
QA-Review: Derek Hansen <dhansen@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>