Unfragile spec failures due to help dialog
The help dialog currently sets an instance variable when init-ed, and
it's kind of being used as a global singleton.  This means that if
it's initialized, used, and then something else wipes the dom, the
next time it gets engaged, it thinks it's already been initialized
and then tries to call methods that the DOM isn't ready for.

Lesson:  you can't count on some global semaphore to be a proxy
for DOM state

Change-Id: Ic5f0216fe9341ada85c30545a894a1a500f09b52
Reviewed-on: https://gerrit.instructure.com/52341
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>