spec: fix fragile gradebook spec
One of these specs was failing because the previous spec would click a
cell, leaving the pointer over that cell when the next spec runs. That
spec would then get the contents of the cell and be confused by the
presence of the tooltip text caused by the hover. This change makes the
check more specific, getting right to the content we're interested in
and avoiding the tooltip.

Change-Id: Iee35c7bf39c45812a26020c8bb59d209d95d6c43
Reviewed-on: https://gerrit.instructure.com/53146
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
Product-Review: Zach Wily <zach@instructure.com>
QA-Review: Zach Wily <zach@instructure.com>