spec: ensure we wait for a new page load when doing a get
This is another attempt to fix selenium spec reliability. Without this
change, when doing a `get` on a URL, we weren't waiting for the new page
to actually load before doing our checks. This leads to stale element
errors, etc.

We do have to be sure to *not* wait for a new page load if the target
URL is the same as the current URL, but with a different hash.

Running the following command:

```
for i in $(seq 10); do
  echo "**** $i"
  bundle exec rspec -e "moving outcomes tree should move a learning outcome group via tree modal" spec/selenium/
done
```

would fail at least half the time without this patchset, and never fails
with it (in my testing).

Change-Id: I2d707f96d7713f69be865355d8a78bad1261d446
Reviewed-on: https://gerrit.instructure.com/53211
Tested-by: Jenkins
Reviewed-by: Zach Wily <zach@instructure.com>
Product-Review: Zach Wily <zach@instructure.com>
QA-Review: Zach Wily <zach@instructure.com>