fix potential m/d/y vs. d/m/y bug
refs CNVS-3194

the $.datepicker.parseDate in canvas is overridden to try parseDatetime
first (on just the value, ignoring the format parameter) before passing
it to the original $.datepicker.parseDate. if parseDatetime then
interpreted it as d/m/y due to locale settings, that would be an error

currently, Date.parse (used by parseDatetime) always uses m/d/y (which
is a bug in itself), so this instance is not an issue. but I can see
that changing in the future (e.g. to fix that bug) and it starting to
cause problems here.

test-plan: N/A

Change-Id: I1b425c7bc8576fd8538b372f32cfd29d4d4ca6f5
Reviewed-on: https://gerrit.instructure.com/53817
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>