[spec] fix failing mark done presenter spec
This fixes a missing part of the pull request merged in with 54ef257
earlier.  I inadvertently left out this portion because it wasn't
actually part of the pull request, but rather a fix that I had put
in manually to correct the failing spec.

Change-Id: I916c37e3afb95fb7f68d2daee45e5f8c2a2506bf
Reviewed-on: https://gerrit.instructure.com/54649
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Product-Review: Dan Minkevitch <dan@instructure.com>
Tested-by: Jenkins
QA-Review: Clay Diffrient <cdiffrient@instructure.com>