remove N+1 queries on term index page
This will make the page a little faster, but unfortunately the bulk of
the time comes from counting active users in each term, and whether we
do that as N queries or 1, the overall time is about the same.  To
really get this page to perform well, we need to look into storing these
counts somewhere that don't have to be calculated every time.

fixes CNVS-21524

test plan:
- set up several enrollment terms with different numbers of courses and
  users in each.
- to go the root account enrollment terms page, and make sure the counts
  of active courses and users is accurate.
- try deleting some of the courses and users and make sure counts update
  appropriately.
- basic regression test of functionality around saving/editing
  enrollment term start/end date overrides

Change-Id: I35fb637ddd91db64c4a5890d58a622f5affa3923
Reviewed-on: https://gerrit.instructure.com/56439
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>