[a11y] Keeps keyboard focus in the + People modal
This makes it so that the third screen of the "Add People" modal
in courses maintains a focus point.  Prior to this commit focus
was lost when proceeding into the third step, eliminating the ability
to tab through to any buttons.  Now focus is set on the "Done" button
allowing tabbing throughout the modal.

closes CNVS-20788

Test Plan:
  - Go to the People section of a course
  - Click the "+ People" button
  - Fill in the steps proceeding through the screens
  - On the last screen, focus should be set to the "Done" button.
    This should allow you to focus through to all the other interactive
    elements of the modal.

Change-Id: Ie3fdfce381c04944c6a1eb5209ef8257ed0eede5
Reviewed-on: https://gerrit.instructure.com/55710
Tested-by: Jenkins
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>