make module unlock date accessible
closes CNVS-19319
refs CNVS-15543

This applies a screenreader-only label to the datetime picker
used in the module unlock UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

TEST PLAN:
- go to create a new module for a course
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the unlock date with no
  other context.
- tab through to the same field and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: Ide2939fc616ea5f7822eca1a6d91c17541f248de
Reviewed-on: https://gerrit.instructure.com/56565
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
Tested-by: Jenkins
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>