make it so anyone can turn "New UI" on
closes: CNVS-21908

Because of the workflow and tooling some QA people
use, they could not test the new UI because they
had to run in "production" but their portals couldn't.

This also means that it is now visible by non site admins.
So schools can see it and turn it on on their own

Change-Id: Ib3a7a5d00ba3bc6a1680a1528898add81dc165f0
Reviewed-on: https://gerrit.instructure.com/56850
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Matt Fairbourn <mfairbourn@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>