(NUI) Cog dropdown being cut off fix
Fixes CNVS-21134

The cog dropdown was being cut off when positioned on
the rightmost side of a div with a #content id because
it had overflow set to hidden.

Test Plan:
-Navigate to the calendar, then to the scheduler. Create
an appointment group. Select the cog dropdown on the
right side of the appointment group. It should not be
cut off.
-Create a course with modules. Set the course home page
as the modules page. Select any of the cog dropdowns on
the any of the modules. None of them should be cut off.
-Navigate to people within a course. Select the cog
dropdown on any person. It should not be cut off.

Change-Id: I5614665eae65d03e6689e5705fdf2806c5390150
Reviewed-on: https://gerrit.instructure.com/57197
Tested-by: Jenkins
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
QA-Review: Adrian Foong <afoong@instructure.com>