add previewing folders in the files section
fixes CNVS-21610

Folders were showing up as a 'file not found' This was because folders
were not handled at all for previewing in the files section. This
commit adds the ability to see a folder preview which is just the name
of the folder with its icon. We could add the ability to click a link
to the folder in the future as a feature request.

Test Plan
Given you're working in new files
And you have a folder created in the root folder
When you select the sub folder
And you click the preview icon in the top toolbar
Then you should see the preview dialog box pop up
And there should be a folder icon in the center
And there should be the name of the folder under the icon

Change-Id: Idd4e527bbc221b8d568c26734d711dce8fc76039
Reviewed-on: https://gerrit.instructure.com/57589
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>